This is a plugin for Compiz 0.8 branch that allows user to have a 3D earth inside the cube, with realtime day and night.
Cube must, of course, be enabled.
Configure using ccsm. You can set the center of the view to your location, specifying longitude and latitude.
Don't forget to set your timezone. Daylight Savings Time is automatic.

Shader will be used if they are supported, enabling multitexturing (texture for the night side) and per-pixel lighting (with specular reflexion on oceans)
libGLEW is required to build the plugin.