Changed the textures for day and night, and some cleanup.
Found a better nightmap with only lights (thus allowing me to paste it directly instead of filtering for the lights).
Now using a real colours day map, the result is much nicer !

By the way, I said a bit of crap in the last commit about float timezone, I just forgot how I did my calculations :-).
But still there are some places that have a not round timezone (such as French Polynesia : GMT-9h30), so the timezone remains a float !