Now a truly spherical skydome, and the sun !
Overriding the original clearing of the screen with the drawing of my own skydome, which is rotated now, though not in the exact position and the sun position.
The skydome image has been horizontaly mirrored to correct its drawing on the inside of a sphere.
The skydome can (and will better be) disabled in the cube plugin, as it will only draw more memory for nothing.
The sun is only a white sphere for the moment, nothing really cool… :)