Realtime clouds, and some lighting tweaks.
Added the possibility to have a realtime cloudmap, downloaded every 3 hour, and an option to disable this to save bandwidth.
If disabled, it will still use the stock cloudmap or the latest downloaded.
Deleted the fragment shader for GLSL 1.10, and modified the original one to only use functions compatible with 1.10, the end result is not changed.
Tweaked the light model so that the earth is brighter, and to roughly simulate atmosphere scattering and have a somewhat more accurate dawn/dusk.