Bug correction in realtime clouds
Jpg file loading in a thread was not safe, thus I had to take it out of the thread.
Also, when reading the jpg, then writing it to the png it was flipped vertically so I took care of it too.

Added a check to see if clouds.jpg was here, to avoid strange behavior on the first start. You can manually trigger the redownloading of the cloudmap by deleting clouds.jpg in ~/.compiz/images